Reorganizar tests: dividir TaskStoreTests.swift en 5 archivos especializados
- TaskStoreBasicTests.swift: Estado inicial, gestión básica de tareas, cálculos
- TaskStoreFormattingTests.swift: Detección de símbolos, formateo, parsing
- TaskStorePersistenceTests.swift: Persistencia local, auto-guardado, ciclo de vida
- TaskStoreTimerTests.swift: Activación de tareas, timer publishers, Combine
- TaskStoreUITests.swift: Texto de resumen, localización, placeholders

Mantiene todos los 63 tests con funcionalidad idéntica.
Mejora la mantenibilidad con archivos más pequeños y organizados por funcionalidad.
